+++ AoE2ScenarioParser/datasets/trigger_lists/decision_option.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from AoE2ScenarioParser.datasets.dataset_enum import _DataSetIntEnums
+
+
+class DecisionOption(_DataSetIntEnums):
+    """
+    This enum class provides the integer values for the Decision Triggered Condition.
+
+    **Examples**:
+
+    >>> DecisionOption.Option1
+    <DecisionOption.Option1: 1>
+    """
+    Option1 = 1
+    Option2 = 2